Address deferred follow-ups (#190)
Closes the three items previously deferred in the PR description.

1. **Extract agentInternalError helper** (mirror of controlErrorResponse)

   New file: `packages/backend/src/routes/agent/helpers.ts` exports
   `agentInternalError(c, err, code, message, logCtx, logMessage)`
   plus the typed `AgentInternalErrorCode` union covering every
   500-tier coarse code the agent surface emits.

   All 8 inline `catch { log + return 500 envelope }` blocks in
   `routes/agent/index.ts` collapse to single-line helper calls. The
   log shape (`{err, ...logCtx}`) is preserved byte-for-byte so the
   existing log-key assertions in `agent-api-contract.test.ts`
   (`hasError`, `taskId`, `severity`, etc.) keep matching.

   Net agent route file: ~683 → ~744 lines (helper call sites are
   slightly larger than the prior 4-line inline blocks because each
   carries explicit code + message + logCtx arguments — but the
   visual repetition is gone and the code-vocabulary is now
   compile-checked via the AgentInternalErrorCode union).

2. **Split openapi/components.ts into per-surface files**

   New directory: `packages/backend/src/openapi/components/` with
   `shared.ts` (ResponseConfig + guardDuplicateComponentRegistration),
   `dashboard.ts`, `control.ts`, `agent.ts`. Each per-surface file
   owns its envelope schema, response names, brand-typed ref union,
   description table, defaultHook, shared-response wrapper, and
   registrar.

   `openapi/components.ts` is now a 54-line barrel that re-exports
   the public API surface. The ~30 importing route files don't need
   to change — the import path
   `'../../openapi/components.js'` resolves to the barrel which
   forwards to the right per-surface file.

3. **Fix silent data-loss in updateTaskProgress**

   `packages/backend/src/services/tasks.ts:594-604` previously logged
   the "campaign has no associated hash list" condition and fell
   through to a successful `{task: updated}` return. The agent saw
   200 + acknowledged and assumed its cracked hashes were stored;
   they were silently dropped. The campaign aggregate would then
   tick forward on phantom progress.

   The corrected branch returns
   `{error: 'Campaign has no associated hash list; cracked results
   cannot be stored. Check campaign configuration before resubmitting.'}`.
   The route handler's existing `'error' in result` branch surfaces
   this as 400 + `TASK_ERROR` so the agent can quarantine the report
   instead of marking the task done with phantom progress.

   New contract test pins the end-to-end wire behavior: forcing the
   service to return the new `{error}` shape produces a 400 with
   `error.code === 'TASK_ERROR'` and a message matching
   `/no associated hash list/i`.

`just ci-check` green: 613 backend tests pass (was 612, 1 new),
5/5 e2e pass.
